tests/requirements-test.txt
# 测试依赖
pytest>=7.0.0
hypothesis>=6.0.0
PyYAML>=6.0
